feat(sec-auth): Refactor AuthContext (Relates #21)
