+++ frontend/src/lib/graphql/queries/auth.queries.ts
@@ -1,0 +1,34 @@
+import { gql } from '@apollo/client';
+
+export const ME_QUERY = gql`
+  query Me {
+    me {
+      id
+      email
+      name
+      role
+    }
+  }
+`;
+
+export const USER_QUERY = gql`
+  query User($id: ID!) {
+    user(id: $id) {
+      id
+      email
+      name
+      role
+    }
+  }
+`;
+
+export const USERS_QUERY = gql`
+  query Users {
+    users {
+      id
+      email
+      name
+      role
+    }
+  }
+`;
